+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <!-- Template, # for major version, ## for minor and patch
 
-# 1.0.0 (YYYY-MM-DD)
+# 1.0.0 (2021-05-15)
 ### Added
 *
 ### Changed
@@ -10,6 +10,14 @@
 ### Fixed
 *
 -->
+
+# 1.1.0 (2021-05-15)
+### Added
+* Press any key to exit program (good for windows)
+* SO Notifier it's done
+### Changed
+* Filename changed from `$day $author $code` to `$day [$author] $code`
+
 
 
 ## 0.1.0 (2021-05-05)